move reading capabilities too
and refactor a bit to take `AsRef<Path>` instead of `&str` as argument
