+++ cosmogony/Cargo.toml
@@ -21,7 +21,9 @@ geo = "0.12"
 geo-types = { version = "0.4", features = ["rstar"] }
 serde = {version = "1", features = ["rc"]}
 serde_derive = "1"
+serde_json = "1"
 geojson = "0.15"
 failure = "0.1"
 failure_derive = "0.1"
+flate2 = "1.0"
 osmpbfreader = "0.13"
