+++ cosmogony/src/file_format.rs
@@ -1,4 +1,5 @@
 use failure::Error;
+use std::path::Path;
 
 #[derive(PartialEq, Clone)]
 pub enum OutputFormat {
@@ -16,10 +17,16 @@ static ALL_EXTENTIONS: [(&str, OutputFormat); 4] = [
 ];
 
 impl OutputFormat {
-    pub fn from_filename(filename: &str) -> Result<OutputFormat, Error> {
+    pub fn from_filename(filename: impl AsRef<Path>) -> Result<OutputFormat, Error> {
         ALL_EXTENTIONS
             .iter()
-            .find(|&&(ref e, _)| filename.ends_with(e))
+            .find(|&&(ref e, _)| {
+                filename
+                    .as_ref()
+                    .file_name()
+                    .and_then(|f| f.to_str())
+                    .map_or(false, |f| f.ends_with(e))
+            })
             .map(|&(_, ref f)| f.clone())
             .ok_or_else(|| {
                 let extensions_str = ALL_EXTENTIONS
@@ -30,7 +37,8 @@ impl OutputFormat {
                 failure::err_msg(format!(
                     "Unable to detect the file format from filename '{}'. \
                      Accepted extensions are: {}",
-                    filename, extensions_str
+                    filename.as_ref().display(),
+                    extensions_str
                 ))
             })
     }
