+++ cosmogony/src/lib.rs
@@ -1,6 +1,9 @@
+pub mod file_format;
 mod model;
 pub mod mutable_slice;
+mod read;
 mod zone;
 
-pub use crate::zone::{Coord, Zone, ZoneIndex, ZoneType};
 pub use model::{Cosmogony, CosmogonyMetadata, CosmogonyStats};
+pub use read::{load_cosmogony_from_file, read_zones_from_file};
+pub use zone::{Coord, Zone, ZoneIndex, ZoneType};
