+++ src/bin/cosmogony.rs
@@ -1,5 +1,5 @@
-use cosmogony::Cosmogony;
-use cosmogony_builder::{build_cosmogony, file_format::OutputFormat};
+use cosmogony::{file_format::OutputFormat, Cosmogony};
+use cosmogony_builder::build_cosmogony;
 use flate2::write::GzEncoder;
 use flate2::Compression;
 use std::fs::File;
